image-classifier/requirements.txt
django
djangorestframework
django-cors-headers
keras